updates for argoneut

+++ buildScripts/buildArgoNeuT.sh
@@ -10,28 +10,54 @@ echo "base qualifiers: $QUAL"
 echo "build type: $BUILDTYPE"
 echo "workspace: $WORKSPACE"
 
-ncores=`cat /proc/cpuinfo 2>/dev/null | grep -c -e '^processor'`
+if [ `uname` = Darwin ]; then
+  #ncores=`sysctl -n hw.ncpu`
+  #ncores=$(( $ncores / 4 ))
+  ncores=16
+else
+  ncores=`cat /proc/cpuinfo 2>/dev/null | grep -c -e '^processor'`
+fi
+if [ $ncores -lt 1 ]; then
+  ncores=1
+fi
+echo "Building using $ncores cores."
 
-source /grid/fermiapp/products/larsoft/setup || exit 1
+source /grid/fermiapp/products/argoneut/setup_argoneut.sh || exit 1
 
-setup git || exit 1
+if ! uname | grep -q Darwin; then
+  setup git || exit 1
+fi
 setup gitflow || exit 1
-setup mrb || exit 1
 export MRB_PROJECT=argoneut
+echo "Mrb path:"
 which mrb
 
-set -x
+#set -x
 rm -rf $WORKSPACE/temp || exit 1
 mkdir -p $WORKSPACE/temp || exit 1
 mkdir -p $WORKSPACE/copyBack || exit 1
 rm -f $WORKSPACE/copyBack/* || exit 1
 cd $WORKSPACE/temp || exit 1
 mrb newDev  -v $ARGONEUTVER -q $QUAL:$BUILDTYPE || exit 1
-set +x
+#set +x
 
 source localProducts*/setup || exit 1
 
-set -x
+# some shenanigans so we can use getopt v1_1_6
+if [ `uname` = Darwin ]; then
+#  cd $MRB_INSTALL
+#  curl --fail --silent --location --insecure -O http://scisoft.fnal.gov/scisoft/packages/getopt/v1_1_6/getopt-1.1.6-d13-x86_64.tar.bz2 || \
+#      { cat 1>&2 <<EOF
+#ERROR: pull of http://scisoft.fnal.gov/scisoft/packages/getopt/v1_1_6/getopt-1.1.6-d13-x86_64.tar.bz2 failed
+#EOF
+#        exit 1
+#      }
+#  tar xf getopt-1.1.6-d13-x86_64.tar.bz2 || exit 1
+  setup getopt v1_1_6  || exit 1
+#  which getopt
+fi
+
+#set -x
 cd $MRB_SOURCE  || exit 1
 # make sure we get a read-only copy
 mrb g -r -t $ARGONEUTVER argoneutcode || exit 1
